[Core] Move options into `io.cucumber.core.options` package
