+++ core/src/main/java/io/cucumber/core/options/EnvironmentOptionsParser.java
@@ -1,5 +1,7 @@
-package cucumber.runtime;
+package io.cucumber.core.options;
 
+import cucumber.runtime.Env;
+import cucumber.runtime.Shellwords;
 import cucumber.runtime.io.MultiLoader;
 import cucumber.runtime.io.ResourceLoader;
 import io.cucumber.core.model.RerunLoader;
