+++ core/src/main/java/io/cucumber/core/options/RuntimeOptionsBuilder.java
@@ -1,6 +1,7 @@
-package cucumber.runtime;
+package io.cucumber.core.options;
 
 import cucumber.api.SnippetType;
+import cucumber.runtime.CucumberException;
 import cucumber.runtime.formatter.PluginFactory;
 import cucumber.runtime.order.PickleOrder;
 import io.cucumber.core.model.FeatureWithLines;
