+++ core/src/main/resources/io/cucumber/core/options/USAGE.txt
