+++ core/src/test/java/cucumber/runtime/BackendModuleBackendSupplierTest.java
@@ -3,6 +3,7 @@ package cucumber.runtime;
 import cucumber.runtime.io.MultiLoader;
 import cucumber.runtime.io.ResourceLoader;
 import cucumber.runtime.io.ResourceLoaderClassFinder;
+import io.cucumber.core.options.RuntimeOptions;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
