+++ core/src/test/java/cucumber/runtime/formatter/JSONFormatterTest.java
@@ -6,10 +6,10 @@ import cucumber.runner.TestBackendSupplier;
 import cucumber.runner.TestHelper;
 import cucumber.runner.TimeServiceEventBus;
 import cucumber.runner.TimeServiceStub;
-import cucumber.runtime.CommandlineOptionsParser;
+import io.cucumber.core.options.CommandlineOptionsParser;
 import cucumber.runtime.HookDefinition;
 import cucumber.runtime.Runtime;
-import cucumber.runtime.RuntimeOptions;
+import io.cucumber.core.options.RuntimeOptions;
 import cucumber.runtime.io.ClasspathResourceLoader;
 import cucumber.runtime.model.CucumberFeature;
 import cucumber.runtime.snippets.FunctionNameGenerator;
