+++ core/src/test/java/cucumber/runtime/formatter/PluginsTest.java
@@ -8,8 +8,8 @@ import cucumber.api.event.EventPublisher;
 import cucumber.api.formatter.ColorAware;
 import cucumber.api.formatter.StrictAware;
 import cucumber.runner.CanonicalOrderEventPublisher;
-import cucumber.runtime.RuntimeOptions;
-import cucumber.runtime.RuntimeOptionsBuilder;
+import io.cucumber.core.options.RuntimeOptions;
+import io.cucumber.core.options.RuntimeOptionsBuilder;
 import org.junit.Rule;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
