+++ core/src/test/java/io/cucumber/core/options/RuntimeOptionsTest.java
@@ -1,4 +1,4 @@
-package cucumber.runtime;
+package io.cucumber.core.options;
 
 import cucumber.api.Plugin;
 import cucumber.api.SnippetType;
@@ -8,6 +8,9 @@ import cucumber.api.formatter.ColorAware;
 import cucumber.api.formatter.StrictAware;
 import cucumber.runner.TimeService;
 import cucumber.runner.TimeServiceEventBus;
+import cucumber.runtime.CucumberException;
+import cucumber.runtime.Env;
+import cucumber.runtime.Shellwords;
 import cucumber.runtime.formatter.PluginFactory;
 import cucumber.runtime.formatter.Plugins;
 import cucumber.runtime.io.Resource;
