+++ junit/src/main/java/io/cucumber/junit/Cucumber.java
@@ -11,10 +11,10 @@ import cucumber.runtime.BackendModuleBackendSupplier;
 import cucumber.runtime.BackendSupplier;
 import cucumber.runtime.ClassFinder;
 import cucumber.runtime.Env;
-import cucumber.runtime.EnvironmentOptionsParser;
+import io.cucumber.core.options.EnvironmentOptionsParser;
 import cucumber.runtime.FeaturePathFeatureSupplier;
-import cucumber.runtime.RuntimeOptions;
-import cucumber.runtime.CucumberOptionsAnnotationParser;
+import io.cucumber.core.options.RuntimeOptions;
+import io.cucumber.core.options.CucumberOptionsAnnotationParser;
 import cucumber.runtime.filter.Filters;
 import cucumber.runtime.formatter.PluginFactory;
 import cucumber.runtime.formatter.Plugins;
